+++ Prototyping/Library/include/Prototyping/UI/Qt/QtTileGridIsometricRenderer.h
@@ -73,7 +73,7 @@ namespace Prototyping::UI::Qt {
             polygon << top << right << bottom << left;
             auto center = polygon.boundingRect().center();
 
-            return UIPosition{static_cast<uint32_t>(center.x()), static_cast<uint32_t>(center.y())};
+            return UIPosition{center.x(), center.y()};
         }
     };
 }
